create doc files and rename import
