+++ apps/dashboard/components/settings/UpdateConfiguration.module.css
