+++ packages/typings/src/stat.ts
@@ -2,6 +2,7 @@ interface IStat {
   timestamp: string;
   cpu: number;
   memory: number;
+  heapUsed: number;
   memoryMax: number;
   uptime: number;
   source: {
